# Documentation/api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/transactions/purchases/draft": {
      "post": {
        "operationId": "TransactionsController_createPurchaseDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Purchase draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Supplier or product not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Supplier or product inactive",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create purchase draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/sales/draft": {
      "post": {
        "operationId": "TransactionsController_createSaleDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSaleDraftDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sale draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Customer or product not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Customer or product inactive",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create sale draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/supplier-payments/draft": {
      "post": {
        "operationId": "TransactionsController_createSupplierPaymentDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierPaymentDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Supplier payment draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Supplier or payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Supplier or payment account inactive",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create supplier payment draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/customer-payments/draft": {
      "post": {
        "operationId": "TransactionsController_createCustomerPaymentDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerPaymentDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer payment draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Customer or payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Customer or payment account inactive",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create customer payment draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/supplier-returns/draft": {
      "post": {
        "operationId": "TransactionsController_createSupplierReturnDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupplierReturnDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Supplier return draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Supplier or source line not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Over-return or invalid source line",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create supplier return draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/customer-returns/draft": {
      "post": {
        "operationId": "TransactionsController_createCustomerReturnDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerReturnDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer return draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Customer or source line not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Over-return or invalid source line",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create customer return draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/internal-transfers/draft": {
      "post": {
        "operationId": "TransactionsController_createInternalTransferDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInternalTransferDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Internal transfer draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or same account",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Payment account not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Payment account inactive",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create internal transfer draft",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/adjustments/draft": {
      "post": {
        "operationId": "TransactionsController_createAdjustmentDraft",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdjustmentDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Adjustment draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "422": {
            "description": "Product inactive",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create adjustment draft (OWNER/ADMIN only)",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/allocations": {
      "get": {
        "operationId": "TransactionsController_listAllocations",
        "parameters": [
          {
            "name": "supplierId",
            "required": false,
            "in": "query",
            "description": "Filter by supplier",
            "schema": { "type": "string" }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Filter by customer",
            "schema": { "type": "string" }
          },
          {
            "name": "purchaseId",
            "required": false,
            "in": "query",
            "description": "Filter by applies-to purchase transaction",
            "schema": { "type": "string" }
          },
          {
            "name": "saleId",
            "required": false,
            "in": "query",
            "description": "Filter by applies-to sale transaction",
            "schema": { "type": "string" }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Filter allocations from this date (inclusive)",
            "schema": { "type": "string" }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Filter allocations to this date (inclusive)",
            "schema": { "type": "string" }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": { "default": 1, "type": "number" }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": { "maximum": 100, "default": 20, "type": "number" }
          }
        ],
        "responses": {
          "200": { "description": "Allocation list" },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List allocations",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions/{id}/post": {
      "post": {
        "operationId": "TransactionsController_post",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Transaction UUID",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PostTransactionDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction posted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Post transaction (finalize & create entries)",
        "tags": ["Transactions"]
      }
    },
    "/api/v1/transactions": {
      "get": {
        "operationId": "TransactionsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["PURCHASE", "SALE", "RETURN", "TRANSFER"],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": { "enum": ["DRAFT", "POSTED", "VOID"], "type": "string" }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "schema": { "example": "2026-02-01", "type": "string" }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "schema": { "example": "2026-02-11", "type": "string" }
          },
          {
            "name": "supplierId",
            "required": false,
            "in": "query",
            "schema": { "example": "supplier-uuid", "type": "string" }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": { "example": "customer-uuid", "type": "string" }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "enum": ["transactionDate", "createdAt", "totalAmount"],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": { "enum": ["asc", "desc"], "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiErrorResponse" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "List transactions",
        "tags": ["Transactions"]
      }
    },
...
  },
  "info": {
    "title": "Persona Finance API",
    "description": "Multi-tenant finance system. Standard error response format includes: `statusCode`, `message`, `errors[]`, `timestamp`, `path`, `requestId`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header"
      }
    },
    "schemas": {
...
      "ApiErrorField": {
        "type": "object",
        "properties": {
          "field": { "type": "string", "example": "name" },
          "message": { "type": "string", "example": "must not be empty" }
        },
        "required": ["field", "message"]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "message": { "type": "string", "example": "Validation failed" },
          "errors": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApiErrorField" }
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-11T12:00:00.000Z"
          },
          "path": { "type": "string", "example": "/api/v1/customers" },
          "requestId": { "type": "string", "example": "req_123" }
        },
        "required": ["statusCode", "message", "timestamp", "path"]
      },
...
      "CreatePurchaseDraftDto": {
        "type": "object",
        "properties": {
          "supplierId": { "type": "string", "example": "supplier-uuid" },
          "transactionDate": { "type": "string", "example": "2026-02-10" },
          "lines": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PurchaseLineDto" }
          },
          "deliveryFee": {
            "type": "number",
            "example": 2000,
            "description": "Delivery fee in PKR (integer)"
          },
          "notes": {
            "type": "string",
            "example": "Handle with care",
            "maxLength": 1000
          },
          "idempotencyKey": {
            "type": "string",
            "example": "client-generated-uuid-v4",
            "maxLength": 64,
            "description": "Client-supplied idempotency key"
          }
        },
        "required": ["supplierId", "transactionDate", "lines"]
      },
...
      "TransactionListResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TransactionResponseDto" }
          },
          "meta": { "$ref": "#/components/schemas/TransactionListMetaDto" }
        },
        "required": ["data", "meta"]
      },
...
    }
  }
}